+++ TombEngine/Objects/TR4/Object/FireRope.h
@@ -1,0 +1,9 @@
+#pragma once
+
+class Vector3i;
+struct CollisionInfo;
+struct ItemInfo;
+
+void TriggerRopeFlame(Vector3i pos);
+void FireRopeControl(short itemNumber);
+void FireRopeCollision(short itemNumber, ItemInfo* laraItem, CollisionInfo* coll);